+++ bin/my-reservations.js
@@ -23,7 +23,7 @@ winston.loggers.add('gomngr', {
 
 const logger = winston.loggers.get('gomngr');
 
-const tps = require('../routes/tp.js');
+const tpssrv = require('../core/tps.service.js');
 
 if (!console.table){
     require('console.table');
@@ -43,7 +43,7 @@ var processReservation = function(reservation){
     // eslint-disable-next-line no-unused-vars
     return new Promise(function (resolve, reject){
         logger.info('create user for reservation ', reservation);
-        tps.exec_tp_reservation(reservation._id, 'auto').then(function(res){
+        tpssrv.exec_tp_reservation(reservation._id, 'auto').then(function(res){
             logger.debug('set reservation as done', res);
             dbsrv.mongo_reservations().updateOne({'_id': res._id},{'$set': {'created': true}}).then(function(){
                 resolve(res);
@@ -114,7 +114,7 @@ function removeReservations(rid) {
             Promise.all(reservation.accounts.map(function(user){
                 return dbsrv.mongo_users().findOne({'uid': user});
             })).then(function(users){
-                return tps.delete_tp_users(users, reservation.group, 'auto');
+                return tpssrv.delete_tp_users(users, reservation.group, 'auto');
             }).then(function(){
                 console.log('[INFO] close reservation', reservations);
                 Promise.all(reservations.map(function(reservation){
